BACKPORT: fs: limit filesystem stacking depth
BACKPORT-NOTE: fs/overlayfs/super.c did not exist in this kernel version.

Add a simple read-only counter to super_block that indicates how deep this
is in the stack of filesystems.  Previously ecryptfs was the only stackable
filesystem and it explicitly disallowed multiple layers of itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
(cherry picked from commit 69c433ed2ecd2d3264efd7afec4439524b319121)
Signed-off-by: Bernie Thompson <bhthompson@google.com>

Change-Id: I92e915a3f75cb8e40b32d9f3cbdf89e571617f85
Reviewed-on: https://chromium-review.googlesource.com/469193
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>