BACKPORT: staging: android: Replace timeval with timespec64
'timeval' is used to print timestamps in seconds and millisecond.

32-bit systems using 'struct timeval' will break in the year 2038,
So we have to replace that code with more appropriate types.

This patch changes the android driver to use timespec64.
Since this is a staging driver and the output is only used for a
debugfs file, it is fine to slightly change the output and
use nanoseconds instead. ktime_to_timespec64, is used which will return
seconds and nanoseconds.

Backport: account for OOO change for creating fences from dma fences.

BUG=chromium:911825
TEST=boot caroline, chell, elm, run login_LoginSuccess, graphics_SanAngeles, cheets_StartAndroid

Signed-off-by: Tapasweni Pathak <tapaswenipathak@gmail.com>
Reviewed-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 353fdf17060b15d1d857899cd3ed9d5db4818acf)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I7611218b469c8a89c293893e3db29cf1baae0d4b