UPSTREAM: r8152: clear SELECTIVE_SUSPEND when autoresuming
The flag of SELECTIVE_SUSPEND should be cleared when autoresuming.
Otherwise, when the system suspend and resume occur, it may have
the wrong flow.

Besides, because the flag of SELECTIVE_SUSPEND couldn't be used
to check if the hw enables the relative feature, it should alwayes
be disabled in close().

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 923e1ee3ff0b585cc4f56cf696c8455708537ffb)

BUG=chrome-os-partner:622013
TEST=whirlwind: run iperf (NAT LAN/WAN)
    samus: confirm http://stanford.speedtest.googlefiber.net/ works

Change-Id: I6dfac1ad6329aa635039330d66f46600344a218c
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/360366
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>