UPSTREAM: kgdb: timeout if secondary CPUs ignore the roundup
Currently if an active CPU fails to respond to a roundup request the CPU
that requested the roundup will become stuck.  This needlessly reduces the
robustness of the debugger.

This patch introduces a timeout allowing the system state to be examined
even when the system contains unresponsive processors.  It also modifies
kdb's cpu command to make it censor attempts to switch to unresponsive
processors and to report their state as (D)ead.

BUG=chrome-os-partner:50495
TEST=Use kgdb / kdb

Change-Id: I9f381dae6365840bc4776e7f6a12ac33251a3c2d
Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>
Cc: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit a1465d2f396e416a0049332b20fca5977384b9f5)
Reviewed-on: https://chromium-review.googlesource.com/329029
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>