UPSTREAM: userns: Check euid no fsuid when establishing an unprivileged uid mapping
setresuid allows the euid to be set to any of uid, euid, suid, and
fsuid.  Therefor it is safe to allow an unprivileged user to map
their euid and use CAP_SETUID privileged with exactly that uid,
as no new credentials can be obtained.

I can not find a combination of existing system calls that allows setting
uid, euid, suid, and fsuid from the fsuid making the previous use
of fsuid for allowing unprivileged mappings a bug.

This is part of a fix for CVE-2014-8989.

BUG=chromium:762903
TEST=Build and run

Change-Id: I1bdf5653654deda160581d1b6ce21c5e0cf0ff4d
Cc: stable@vger.kernel.org
Reviewed-by: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 80dd00a23784b384ccea049bfb3f259d3f973b9d)
Reviewed-on: https://chromium-review.googlesource.com/660816
Reviewed-by: Dylan Reid <dgreid@chromium.org>