UPSTREAM: dm thin: fix missing out-of-data-space to write mode transition if blocks are released
commit 2c43fd26e46734430122b8d2ad3024bb532df3ef upstream.

Discard bios and thin device deletion have the potential to release data
blocks.  If the thin-pool is in out-of-data-space mode, and blocks were
released, transition the thin-pool back to full write mode.

The correct time to do this is just after the thin-pool metadata commit.
It cannot be done before the commit because the space maps will not
allow immediate reuse of the data blocks in case there's a rollback
following power failure.

BUG=chromium:616822
TEST=build/boot on samus, xfstest generic/347 passes

Change-Id: I1f46167ceebe673b2a1a3f7bfa74d5f552c1c962
Signed-off-by: Joe Thornber <ejt@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry-picked from 1b210c873eba from 3.14-stable)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/699869
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>