UPSTREAM: netlink/genetlink: pass network namespace to bind/unbind
Netlink families can exist in multiple namespaces, and for the most
part multicast subscriptions are per network namespace. Thus it only
makes sense to have bind/unbind notifications per network namespace.

To achieve this, pass the network namespace of a given client socket
to the bind/unbind functions.

Also do this in generic netlink, and there also make sure that any
bind for multicast groups that only exist in init_net is rejected.
This isn't really a problem if it is accepted since a client in a
different namespace will never receive any notifications from such
a group, but it can confuse the family if not rejected (it's also
possible to silently (without telling the family) accept it, but it
would also have to be ignored on unbind so families that take any
kind of action on bind/unbind won't do unnecessary work for invalid
clients like that.

BUG=706213
TEST=Build and boot kernel

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 023e2cfa36c31b0ad28c159a1bb0d61ff57334c8)
Signed-off-by: Abhishek Bhardwaj <abhishekbh@google.com>

Change-Id: Ic6093f2ead5fab9744c629612777c6fc8b19843e
Reviewed-on: https://chromium-review.googlesource.com/464188
Commit-Ready: Abhishek Bhardwaj <abhishekbh@google.com>
Tested-by: Abhishek Bhardwaj <abhishekbh@google.com>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>