UPSTREAM: r8152: remove generic_ocp_read before writing
For ocp_write_word() and ocp_write_byte(), there is a generic_ocp_read()
which is used to read the whole 4 byte data, keep the unchanged bytes,
and modify the expected bytes. However, the "byen" could be used to
determine which bytes of the 4 bytes to write, so the action could be
removed.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 8cb3db24c80958647f8a1b65e4968f9f043ffc3d)

BUG=chrome-os-partner:622013
TEST=built/installed on gale and sentry. iperf no longer TX timeouts.

Change-Id: I1cb92a3f7d7ef2b651f8bf003c543b12fe20cc7f
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/360384
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>