ANDROID: sched: Determine the current sched_group idle-state
To estimate the energy consumption of a sched_group in
sched_group_energy() it is necessary to know which idle-state the group
is in when it is idle. For now, it is assumed that this is the current
idle-state (though it might be wrong). Based on the individual cpu
idle-states group_idle_state() finds the group idle-state.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

BUG=chrome-os-partner:58667
TEST=build/boot on kevin

Change-Id: I70b61495470700b1d0e87cf92d063b17e7644118
Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
(cherry picked from commit ec055b978e851118a2b9d42f61ecad619cd1ce7
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/420007
(cherry picked from commit 5fcc33d40815b3a454857c4d73a3d79307a6aac2)
Reviewed-on: https://chromium-review.googlesource.com/423206