FROMGIT: ext4: ignore journal checksum on remount; don't fail
commit 2d5b86e048780c5efa7f7d9708815555919e7b05 upstream.

As of v3.18, ext4 started rejecting a remount which changes the
journal_checksum option.

Prior to that, it was simply ignored; the problem here is that
if someone has this in their fstab for the root fs, now the box
fails to boot properly, because remount of root with the new options
will fail, and the box proceeds with a readonly root.

I think it is a little nicer behavior to accept the option, but
warn that it's being ignored, rather than failing the mount,
but that might be a subjective matter...

Reported-by: Cónräd <conradsand.arma@gmail.com>
Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: Josh Boyer <jwboyer@fedoraproject.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 4e55dad93ed97ddcce8f4517683cf9aafa22337b in
 git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git linux-3.18.y)

BUG=chromium:948948
TEST=Run xfstests. generic/{92, 286, 338} pass with entire tree.

Cq-Depend: chromium:1807926
Change-Id: Ifb7b4c6e6d8ccc7120b199f676e3b3a17a18e7f5
Signed-off-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1807893
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Ready: Sarthak Kukreti <sarthakkukreti@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>