BACKPORT: sg_start_req(): make sure that there's not too many elements in iovec
unfortunately, allowing an arbitrary 16bit value means a possibility of
overflow in the calculation of total number of pages in bio_map_user_iov() -
we rely on there being no more than PAGE_SIZE members of sum in the
first loop there.  If that sum wraps around, we end up allocating
too small array of pointers to pages and it's easy to overflow it in
the second loop.

BUG=chromium:669439
TEST=Build and Compile

Change-Id: Ic81dd7bf54b1c6d41d890e39d8563c5da2ac48fe
X-Coverup: TINC (and there's no lumber cartel either)
Cc: stable@vger.kernel.org # way, way back
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
(cherry picked from commit 451a2886b6b)
[groeck: Upstream used wrong (non-existing) definition MAX_UIOVEC]
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/415116
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>